Removing LinearMath C++ Header
