+++ test_tf2/test/test_message_filter.cpp
@@ -41,8 +41,8 @@
 #include <geometry_msgs/msg/transform_stamped.hpp>
 #include <rclcpp/rclcpp.hpp>
 #include <tf2/buffer_core.hpp>
-#include <tf2/LinearMath/Quaternion.hpp>
-#include <tf2/LinearMath/Vector3.hpp>
+#include <tf2/LinearMath/Quaternion.h>
+#include <tf2/LinearMath/Vector3.h>
 #include <tf2/time.hpp>
 #include <tf2_ros/buffer_interface.h>
 #include <tf2_ros/create_timer_ros.h>
