+++ test_tf2/test/test_utils.cpp
@@ -18,7 +18,7 @@
 #include <geometry_msgs/msg/quaternion_stamped.hpp>
 #include <geometry_msgs/msg/transform.hpp>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
-#include <tf2/LinearMath/Quaternion.hpp>
+#include <tf2/LinearMath/Quaternion.h>
 #include <tf2/utils.hpp>
 #include <tf2_kdl/tf2_kdl.hpp>
