+++ tf2/CMakeLists.txt
@@ -62,13 +62,6 @@ if(BUILD_TESTING)
     include/tf2/LinearMath/Scalar.h
     include/tf2/LinearMath/Transform.h
     include/tf2/LinearMath/Vector3.h
-    include/tf2/LinearMath/Matrix3x3.hpp
-    include/tf2/LinearMath/MinMax.hpp
-    include/tf2/LinearMath/QuadWord.hpp
-    include/tf2/LinearMath/Quaternion.hpp
-    include/tf2/LinearMath/Scalar.hpp
-    include/tf2/LinearMath/Transform.hpp
-    include/tf2/LinearMath/Vector3.hpp
   )
 
   ament_copyright(EXCLUDE ${_linter_excludes})
