+++ tf2/include/tf2/LinearMath/MinMax.h
@@ -13,11 +13,59 @@ subject to the following restrictions:
 */
 
 
-#ifndef TF2__LINEARMATH__MINMAX_H_
-#define TF2__LINEARMATH__MINMAX_H_
 
-#warning This header is obsolete, please include tf2/LinearMath/MinMax.hpp instead
+#ifndef GEN_MINMAX_H
+#define GEN_MINMAX_H
 
-#include <tf2/LinearMath/MinMax.hpp>
+#include "Scalar.h"
 
-#endif  // TF2__LINEARMATH__MINMAX_H_
+template <class T>
+TF2SIMD_FORCE_INLINE const T& tf2Min(const T& a, const T& b)
+{
+  return a < b ? a : b ;
+}
+
+template <class T>
+TF2SIMD_FORCE_INLINE const T& tf2Max(const T& a, const T& b)
+{
+  return  a > b ? a : b;
+}
+
+template <class T>
+TF2SIMD_FORCE_INLINE const T& GEN_clamped(const T& a, const T& lb, const T& ub)
+{
+	return a < lb ? lb : (ub < a ? ub : a);
+}
+
+template <class T>
+TF2SIMD_FORCE_INLINE void tf2SetMin(T& a, const T& b)
+{
+    if (b < a)
+	{
+		a = b;
+	}
+}
+
+template <class T>
+TF2SIMD_FORCE_INLINE void tf2SetMax(T& a, const T& b)
+{
+    if (a < b)
+	{
+		a = b;
+	}
+}
+
+template <class T>
+TF2SIMD_FORCE_INLINE void GEN_clamp(T& a, const T& lb, const T& ub)
+{
+	if (a < lb)
+	{
+		a = lb;
+	}
+	else if (ub < a)
+	{
+		a = ub;
+	}
+}
+
+#endif
