+++ tf2/include/tf2/transform_storage.hpp
@@ -30,8 +30,8 @@
 #ifndef TF2__TRANSFORM_STORAGE_HPP_
 #define TF2__TRANSFORM_STORAGE_HPP_
 
-#include "tf2/LinearMath/Vector3.hpp"
-#include "tf2/LinearMath/Quaternion.hpp"
+#include "tf2/LinearMath/Vector3.h"
+#include "tf2/LinearMath/Quaternion.h"
 #include "tf2/time.hpp"
 #include "tf2/visibility_control.hpp"
