+++ tf2/src/cache.cpp
@@ -38,9 +38,9 @@
 #include "tf2/time_cache.hpp"
 #include "tf2/exceptions.hpp"
 
-#include "tf2/LinearMath/Vector3.hpp"
-#include "tf2/LinearMath/Quaternion.hpp"
-#include "tf2/LinearMath/Transform.hpp"
+#include "tf2/LinearMath/Vector3.h"
+#include "tf2/LinearMath/Quaternion.h"
+#include "tf2/LinearMath/Transform.h"
 
 namespace tf2
 {
