+++ tf2/src/static_cache.cpp
@@ -34,7 +34,7 @@
 #include "tf2/time_cache.hpp"
 #include "tf2/exceptions.hpp"
 
-#include "tf2/LinearMath/Transform.hpp"
+#include "tf2/LinearMath/Transform.h"
 
 bool tf2::StaticCache::getData(
   tf2::TimePoint time,
